+++ lib/src/mapbox_maps_platform.dart
@@ -25,6 +25,7 @@ class _MapboxMapsPlatform {
       ArgumentCallbacks<StyleImageMissingEventData>();
   final onStyleImageUnusedPlatform =
       ArgumentCallbacks<StyleImageUnusedEventData>();
+  final onResourceRequestPlatform = ArgumentCallbacks<ResourceEventData>();
 
   final int _channelSuffix = _suffixesRegistry.getSuffix();
   late MethodChannel _channel;
@@ -97,6 +98,10 @@ class _MapboxMapsPlatform {
       case _MapEvent.styleImageRemoveUnused:
         onStyleImageUnusedPlatform(
             StyleImageUnusedEventData.fromJson(jsonDecode(call.arguments)));
+        break;
+      case _MapEvent.resourceRequest:
+        onResourceRequestPlatform(
+            ResourceEventData.fromJson(jsonDecode(call.arguments)));
         break;
       default:
         throw MissingPluginException();